Create README.md
Input data is produced by inputting information related to themselves by the user. Recommendation processing of the data obtained for consumption using kafka as a channel, under the condition that it is based on the data about the professor in the database. Recommendation results are stored back in the database.